SpliceHack Lore: Part One.
For a long time, I have wanted to make Splice more cohesive. Part of that is justifying why this dungeon, with all of its dinosaurs and weirdness, exists. The story will probably never be a big part of Splice, but it will be there for the players that search for it.